CMDB/Asset/requirements.txt
configparser==3.7.3
netifaces==0.10.9
numpy==1.16.2
pandas==0.24.2
pika==0.13.1
psutil==5.6.1
py-dmidecode==0.0.2
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
docker-py==1.10.6
docker-pycreds==0.4.0
